requirements.txt
python-binance
configparser
argparser